Finished.

+++ Task1/Task1/Task1/Car/Car.cs
@@ -13,7 +13,7 @@ namespace Task1Library.Car
         /// <summary>
         /// Represents maximum reachable speed of the car.
         /// </summary>
-        public double MaxSpeed => Engine.Power * 1000 / Weight; // Magic formula to imitate variability of the speed.
+        public int MaxSpeed => (int)(Engine.Power * 1000 / Weight); // Magic formula to imitate variability of the speed.
         
         /// <summary>
         /// The number of car seats.
